Added configuration option to disable the use of the default routes
Unlike Comfortable Mexican Sofa, Comfy-blog loads its built-in routes
by default and there is no way to customize them or the sequence in
which all routes are loaded.
To remedy this a new option 'use_default_routes' is added in
config/initializers/comfy_blog.rb that mimics the behavior of the same
option present in Comfortable Mexican Sofa.